Agents/auth-service-agent/requirements.txt
# Requirements for Auth Service Agent
PyJWT>=2.4.0
bcrypt>=3.2.0